Refactor tests for fewer deps
- should be able to run subset of tests when
  optional dependencies are not installed
- test_bench
  - skip when distributed is not installed
  - skip some tests when stackstac is missing
- Use odc.geo.data instead of geopandas
